+++ tests/testthat/test_inference.R
@@ -8,12 +8,14 @@ d <- data.frame(
 )
 
 test_that("Two-sample tests use appropriate CI", {
-    pTRUE <- getPlotSummary(x, y, data = d,
+    pTRUE <- getPlotSummary(x, y,
+        data = d,
         summary.type = "inference",
         inference.type = "conf",
         hypothesis.var.equal = TRUE
     )
-    pFALSE <- getPlotSummary(x, y, data = d,
+    pFALSE <- getPlotSummary(x, y,
+        data = d,
         summary.type = "inference",
         inference.type = "conf",
         hypothesis.var.equal = FALSE
@@ -21,12 +23,14 @@ test_that("Two-sample tests use appropriate CI", {
 
     pvals <- sapply(list(pTRUE, pFALSE), function(p) {
         as.numeric(
-            gsub(".+=", "",
+            gsub(
+                ".+=", "",
                 strsplit(p[grep("p-value = ", p)[1]], ",")[[1]][3]
             )
         )
     })
-    expect_equal(pvals,
+    expect_equal(
+        pvals,
         c(
             t.test(x ~ y, data = d, var.equal = TRUE)$p.value,
             t.test(x ~ y, data = d)$p.value
@@ -34,10 +38,13 @@ test_that("Two-sample tests use appropriate CI", {
     )
 
     cis <- lapply(list(pTRUE, pFALSE), function(p) {
-        scan(text = gsub("A - B", "", p[grep("A - B", p)]), what = double(),
-            quiet = TRUE)[-1]
+        scan(
+            text = gsub("A - B", "", p[grep("A - B", p)]), what = double(),
+            quiet = TRUE
+        )[-1]
     })
-    expect_equal(cis,
+    expect_equal(
+        cis,
         list(
             as.numeric(round(t.test(x ~ y, data = d, var.equal = TRUE)$conf.int, 3)),
             as.numeric(round(t.test(x ~ y, data = d, var.equal = FALSE)$conf.int, 3))
@@ -58,21 +65,24 @@ ptest <- list(
 )
 btest <- binom.test(table(d$x), p = 0.4, alternative = "less")
 ctest <- chisq.test(table(d$x))
-s1 <- getPlotSummary(x, data = d, summary.type = "inference",
+s1 <- getPlotSummary(x,
+    data = d, summary.type = "inference",
     inference.type = "conf",
     hypothesis.test = "proportion",
     hypothesis.use.exact = FALSE,
     hypothesis.value = 0.5,
     hypothesis.alt = "two.sided"
 )
-s2 <- getPlotSummary(x, data = d, summary.type = "inference",
+s2 <- getPlotSummary(x,
+    data = d, summary.type = "inference",
     inference.type = "conf",
     hypothesis.test = "proportion",
     hypothesis.use.exact = TRUE,
     hypothesis.value = 0.4,
     hypothesis.alt = "less"
 )
-s3 <- getPlotSummary(x, data = d, summary.type = "inference",
+s3 <- getPlotSummary(x,
+    data = d, summary.type = "inference",
     inference.type = "conf",
     hypothesis.test = "chi2"
 )
@@ -126,10 +136,11 @@ d <- expand.grid(
     Machine = c("Desktop of tablet", "Mobile"),
     Course = c("STATS101/G", "STATS108")
 )
-d <- d[rep(1:4, c(4, 3, 1, 2)),]
+d <- d[rep(1:4, c(4, 3, 1, 2)), ]
 rownames(d) <- NULL
 ctest <- chisq.test(table(d$Course, d$Machine), simulate = TRUE)
-s1 <- getPlotSummary(Machine, Course, data = d, summary.type = "inference",
+s1 <- getPlotSummary(Machine, Course,
+    data = d, summary.type = "inference",
     inference.type = "conf",
     hypothesis.test = "chi2"
 )
@@ -143,7 +154,8 @@ test_that("Simulated p-value is included when small expected values", {
 
 test_that("Simulated p-value is included when requested", {
     cas <- read.csv("cas.csv", stringsAsFactors = TRUE)
-    s <- getPlotSummary(cellsource, gender, data = cas,
+    s <- getPlotSummary(cellsource, gender,
+        data = cas,
         summary.type = "inference",
         inference.type = "conf",
         hypothesis.test = "chi2",
@@ -165,26 +177,34 @@ test_that("Simulated p-value is included when requested", {
 test_that("inzinference gives the same output", {
     expect_equal(
         inzinference(Sepal.Length ~ Sepal.Width,
-            data = iris, trend = "linear", width = 80),
+            data = iris, trend = "linear", width = 80
+        ),
         getPlotSummary(Sepal.Width, Sepal.Length,
             data = iris, trend = "linear", width = 80,
-            summary.type = "inference", inference.type = "conf")
+            summary.type = "inference", inference.type = "conf"
+        )
     )
 
     expect_equal(
         inzinference(Sepal.Length ~ Sepal.Width | Species,
-            data = iris, trend = "linear", width = 80),
-        getPlotSummary(Sepal.Width, Sepal.Length, g1 =  Species,
+            data = iris, trend = "linear", width = 80
+        ),
+        getPlotSummary(Sepal.Width, Sepal.Length,
+            g1 = Species,
             data = iris, trend = "linear", width = 80,
-            summary.type = "inference")
+            summary.type = "inference"
+        )
     )
 
     expect_equal(
         inzinference(Sepal.Length ~ Species | Sepal.Width,
-            data = iris, width = 80),
-        getPlotSummary(Sepal.Length, Species, g1 = Sepal.Width,
+            data = iris, width = 80
+        ),
+        getPlotSummary(Sepal.Length, Species,
+            g1 = Sepal.Width,
             data = iris, width = 80, inference.type = "conf",
-            summary.type = "inference")
+            summary.type = "inference"
+        )
     )
 })
 
@@ -198,8 +218,10 @@ test_that("ANOVA (one-way) output is the correct way around", {
 ####
 test_that("Confidence level can be adjusted - dot plots", {
     # dot plot - one
-    inf <- inzinference(~Sepal.Length, data = iris,
-        ci.width = 0.8) |> as.character()
+    inf <- inzinference(~Sepal.Length,
+        data = iris,
+        ci.width = 0.8
+    ) |> as.character()
     expect_match(inf, "Mean with 80% Confidence Interval", all = FALSE)
     expect_equal(
         inf[grep("Estimate\\s+Lower\\s+Upper", inf) + 1L] |>
@@ -209,17 +231,23 @@ test_that("Confidence level can be adjusted - dot plots", {
 
     # dot plot - two
     iris2 <- iris[iris$Species != "setosa", ] |> droplevels()
-    inf <- inzinference(Sepal.Length ~ Species, data = iris2,
-        ci.width = 0.90) |> as.character()
+    inf <- inzinference(Sepal.Length ~ Species,
+        data = iris2,
+        ci.width = 0.90
+    ) |> as.character()
     expect_match(inf, "Group Means with 90% Confidence Intervals",
-        all = FALSE)
-    x <- paste(collapse = "\n",
+        all = FALSE
+    )
+    x <- paste(
+        collapse = "\n",
         inf[grep("Group Means with 90% Confidence Intervals", inf) + 3:4]
     )
     m0 <- read.table(textConnection(x))[, -1] |> as.matrix()
     ci <- tapply(iris2$Sepal.Length, iris2$Species, t.test,
-        conf.level = 0.9) |>
-        sapply(function(x) x$conf.int) |> t()
+        conf.level = 0.9
+    ) |>
+        sapply(function(x) x$conf.int) |>
+        t()
     m <- cbind(
         tapply(iris2$Sepal.Length, iris2$Species, mean),
         ci[, 1],
@@ -238,17 +266,23 @@ test_that("Confidence level can be adjusted - dot plots", {
     )
 
     # dot plot - 3+
-    inf <- inzinference(Sepal.Length ~ Species, data = iris,
-        ci.width = 0.99) |> as.character()
+    inf <- inzinference(Sepal.Length ~ Species,
+        data = iris,
+        ci.width = 0.99
+    ) |> as.character()
     expect_match(inf, "Group Means with 99% Confidence Intervals",
-        all = FALSE)
-    x <- paste(collapse = "\n",
+        all = FALSE
+    )
+    x <- paste(
+        collapse = "\n",
         inf[grep("Group Means with 99% Confidence Intervals", inf) + 3:5]
     )
     m0 <- read.table(textConnection(x))[, -1] |> as.matrix()
     ci <- tapply(iris$Sepal.Length, iris$Species, t.test,
-        conf.level = 0.99) |>
-        sapply(function(x) x$conf.int) |> t()
+        conf.level = 0.99
+    ) |>
+        sapply(function(x) x$conf.int) |>
+        t()
     m <- cbind(
         tapply(iris$Sepal.Length, iris$Species, mean),
         ci[, 1],
@@ -261,43 +295,55 @@ test_that("Confidence level can be adjusted - dot plots", {
         all = FALSE,
         fixed = TRUE
     )
-    x <- paste(collapse = "\n",
+    x <- paste(
+        collapse = "\n",
         gsub("\n", "", inf[grep("adjusted for multiple comparisons", inf) + c(4, 5, 6)],
-            fixed = TRUE)
+            fixed = TRUE
+        )
     )
-    m0 <- read.fwf(textConnection(x), c(31, 9, 10, 10, 10))[, 2:4] |> as.matrix()
+    m0 <- read.fwf(textConnection(x), c(29, 9, 10, 10, 10))[, 2:4] |> as.matrix()
     f <- lm(Sepal.Length ~ Species, data = iris)
     m <- s20x::multipleComp(f, 0.99)[, 1:3]
     expect_equivalent(m0, m)
 
     # bar chart - one way
-    inf <- inzinference(~ Species, data = iris,
-        ci.width = 0.92) |> as.character()
+    inf <- inzinference(~Species,
+        data = iris,
+        ci.width = 0.92
+    ) |> as.character()
     expect_match(
         inf,
         "Estimated Proportions with 92% Confidence Interval",
         all = FALSE
     )
-    x <- paste(collapse = "\n",
+    x <- paste(
+        collapse = "\n",
         inf[grep("Estimated Proportions with 92% Confidence Interval", inf) + 3:5]
     )
-    m0 <- read.table(textConnection(x))[, -1] |> as.matrix() |> unname()
-    t <- qnorm(0.96) * sqrt(1/3 * 2/3 / nrow(iris))
+    m0 <- read.table(textConnection(x))[, -1] |>
+        as.matrix() |>
+        unname()
+    t <- qnorm(0.96) * sqrt(1 / 3 * 2 / 3 / nrow(iris))
     m <- cbind(
-        rep(1/3, 3),
-        1/3 - t,
-        1/3 + t
+        rep(1 / 3, 3),
+        1 / 3 - t,
+        1 / 3 + t
     ) |> round(3)
     expect_equivalent(m0, m)
 
     expect_match(inf, "with 92% Confidence Intervals", all = FALSE)
-    x <- paste(collapse = "\n",
+    x <- paste(
+        collapse = "\n",
         gsub("\n", "", inf[grep("92% Confidence Intervals", inf) + 4:6], fixed = TRUE)
     )
     m0 <- read.fwf(textConnection(x), c(26, 9, 10, 9))[, -1] |>
-        as.matrix() |> as.double() |> round(3)
+        as.matrix() |>
+        as.double() |>
+        round(3)
     m <- freq1way.edited(t(as.matrix(table(iris$Species))), conf.level = 0.92)[, -(1:2)] |>
-        as.matrix() |> as.double() |> round(3)
+        as.matrix() |>
+        as.double() |>
+        round(3)
     expect_equal(m0, m)
 
     # bar chart - two way